refactor(platform): SUG-110 — move roadmap inline to GovernancePage, delete RoadmapPage
Roadmap tables (In Progress + Backlog) now render as a section on
/platform/governance rather than a standalone /platform/roadmap route.
Shipped epics excluded — covered by the existing Recent Releases section.

- GovernancePage: roadmap section wired to stats.linearRoadmap with DataTable,
  priority badges, label chips, stale callout. Stat tiles now live from stats.json.
- RoadmapPage.jsx deleted; /platform/roadmap route removed from App.jsx
- PLATFORM_ROUTES.roadmap removed from routes.js
- PlatformSidebar: Roadmap nav item changed to #roadmap hash anchor on governance
- SUG-110 epic doc updated to reflect inline-on-governance architecture
